Move FileUploadComponent to SupportFileUploads
